Vmm RPC interface exposed by VmmController
This commit mostly moves things around to better isolate areas of
concern.

VmmAction was defined in the `api_server` crate while VmmActionError
was defined in the `Vmm`. The two enums actually define a RPC API
interface to the Vmm.

The VmmController (partly) implements that RPC interface to the
underlying Vmm, so make it also define and expose the RPC API
interface definition.

The Vmm should not try and care about user-facing errors.
The Vmm defines its own set of specific errors which will be later
encapsulated and appended with extra information by the upper layers.

StartMicrovmError is now exposed by the builder, unused errors are
removed.

Some errors tests were updated to work while some of them were
commented out. The plan is to get rid of 'ErrorKind' in a future
commit which will also get rid of many test lines in that area.

Also make a change to the resources store to now own the kernel-file
and cmdline instead of just descriptions of them.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>