api_server_adapter: make the api eventfd blocking
This fixes a race during config/pre-boot phase where the VMM thread
reads this event_fd before API thread gets the chance to increment it.

The API event_fd is used by the API thread to signal the VMM thread
that it has pending messages.

During config/pre-boot phase, this event_fd is used in a simple
blocking loop that waits for an API event, then processes the request.
To keep this scenario simple, make the event_fd a blocking event_fd.

During runtime/post-boot phase the event_fd is registered with polly
alongside multiple other events (from multiple sources), and is only
read when EPOLLIN is reported on it. Having it as a blocking event_fd
is safe.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>