logging.py: open FIFO with flags
`test_sigbus_sigsegv` requires a blocking pipe for the test
to work. Added support for setting specific flags when opening
a FIFO. This way, we can use the FIFO in blocking or non blocking
mode upfront, as well as we can set the flags afterwards, if we
need to change the behavior for the opened pipe.

Signed-off-by: Iulian Barbu <iul@amazon.com>