device.rs: increment metric instead of logging error
We've introduced error log for the case when there is
no available buffer for net RX queue. This can happen for
heavy network traffic usage, and it is not an actual
error, rather an limitation for which we must increment
a specific metric.

Added a similar metric of the block device.

Signed-off-by: iulianbarbu <iul@amazon.com>